first version with petitfs
the sdfat-internal buffer of 512bytes was killing the system.
the application buffer only had ~512 bytes itself for buffering to the DAC.
this was not enough.

therefore now the experiment with petitfs, which does not need an internal buffer.
So now we can use an application buffer of 2*512 Bytes or 4*256 Bytes.

This should be way better than before!

we will see...

Signed-off-by: ludgerknorps <56650955+ludgerknorps@users.noreply.github.com>